Move google charts stubs to own folder
